Consolidate Optional utility to the core library (#20)
* Consolidate Optional utility to the core library

* Fix extra space after file
