+++ rx3/build.gradle
@@ -14,7 +14,6 @@ dependencies {
 
     // RxJava
     api libs.reactivex.rx3
-    api libs.guava
 
     // Test
     testImplementation libs.junit
